insuranceApp/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "insurance-app-backend"
version = "0.1.0"
description = "Insurance App Backend"
authors = [
    {name = "Insurance App Team"}
]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.103.0",
    "uvicorn>=0.23.2",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.3.0",
    "psycopg2-binary>=2.9.7",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.1",
]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 